+++ src/renderer/keybindings.ts
@@ -15,7 +15,6 @@ export enum KeybindAction {
   MessageList_PageUp = 'msglist:pageup',
   MessageList_PageDown = 'msglist:pagedown',
   GlobalGallery_Open = 'globalgallery:open',
-  GlobalMap_Open = 'globalmap:open',
 
   // Actions that are not necessarily triggered by keybindings
   ChatList_SwitchToArchiveView = 'chatlist:switch-to-archive-view',